Infinite recursive call in FunctionExpression.IntFunction.getFloat(t)
reported by ycuicui at https://github.com/prefuse/Prefuse/issues/8

+++ src/prefuse/data/expression/FunctionExpression.java
@@ -179,7 +179,7 @@ abstract class IntFunction extends FunctionExpression {
         return (long)getInt(t);
     }
     public float getFloat(Tuple t) {
-        return (float)getFloat(t);
+        return (float)getInt(t);
     }
     public double getDouble(Tuple t) {
         return (double)getInt(t);
